Add copyValue method to ObjectAccessBarrier
Add new method to MM_ObjectAccessBarrier to copy fields of a value class
instance into another value class instance.

This method differs from the cloneObject method in that the target
instances can be flattened within other objects, arrays,  or values, and is
intended to be used with bytecodes such as putfield, getfield, etc.

Since values may be flattened, it may not have an object header. For
this reason, the method must be given the value's J9Class explicitly.
Flattening also means values may not be 8-byte aligned. In
nocompressedrefs, this means copyValue cannot simply iterate through
fields 8-bytes at a time with the description bits the way cloneObject
does, as mixedObjectReadObject and mixedObjectStoreObject would be
performing unaligned reads and writes.

Signed-off-by: Eric Zhang <eric99.zhang@gmail.com>