Add ValueTypes fields to RAM Class and consistency checking
Add consistency checking between field entries and the valueTypeClasses
ROMClass attribute in loadFlattenableFieldClasses. If a field with the
ACC_FLATTENABLE modifier is not actually a value type (has
J9AccValueType modifier), throw an ICCE. If the field is not in the
valueTypeClasses attribute, throw a CFE. Otherwise, add a reference to
the field J9Class to its owner's J9Class->valueTypes array.

This change also delays loading static value type fields and adding
them to the valueTypes array until class initialization. A flag on
the least significant bit of the J9Class pointer is used to indicate
that that it is loaded and valid. The pointers to the J9UTF8's of the
static field class names are temporarily kept where their J9Class
pointers eventually go.

Signed-off-by: Mohammad Ali Nikseresht <anikser@ibm.com>